% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescale.R
\name{add_geo_timescale}
\alias{add_geo_timescale}
\title{Add geological timescale to a tree plot}
\usage{
add_geo_timescale(
  p,
  tree,
  levels,
  layout,
  version = "ICS 2023/02",
  actual_ntips = NULL,
  timescale_mode = "radial",
  timescale_position = "right",
  angle = 360,
  tree_start_position = "right"
)
}
\arguments{
\item{p}{ggplot object}

\item{tree}{phylo object (edge lengths in Ma)}

\item{levels}{Timescale levels vector, e.g., c("eras", "eons")}

\item{layout}{Layout type}

\item{version}{Geological timescale version. Default: "ICS 2023/02".}
}
\value{
ggplot object
}
\description{
Add geological timescale to a tree plot
}
\keyword{internal}
