% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-smart.R
\name{add_smart_legend}
\alias{add_smart_legend}
\title{Add smart legend to a tree plot}
\usage{
add_smart_legend(p, colors, rank_name, position, nrow, ncol)
}
\arguments{
\item{p}{ggplot object}

\item{colors}{Named color vector}

\item{rank_name}{Display name for the rank}

\item{position}{Position: cardinal direction or length-2 numeric vector}

\item{nrow}{User-specified rows}

\item{ncol}{User-specified columns}
}
\value{
ggplot object
}
\description{
Add smart legend to a tree plot
}
\keyword{internal}
