% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-labels.R
\name{add_support_labels}
\alias{add_support_labels}
\title{Add node support labels to a tree plot}
\usage{
add_support_labels(p, tree, threshold)
}
\arguments{
\item{p}{ggplot object}

\item{tree}{phylo object (with node.data if from treedata)}

\item{threshold}{Minimum support value to display}
}
\value{
ggplot object
}
\description{
Add node support labels to a tree plot
}
\keyword{internal}
