% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{build_plot_timetree_params}
\alias{build_plot_timetree_params}
\title{Build plot_timetree parameter list from CLI options}
\usage{
build_plot_timetree_params(
  opt,
  color_palette,
  color_mapping = NULL,
  taxonomy_levels = NULL
)
}
\description{
Build plot_timetree parameter list from CLI options
}
\keyword{internal}
