% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selftest.R
\name{check_dependency}
\alias{check_dependency}
\title{Check a single dependency}
\usage{
check_dependency(name, min_version, required, verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Report progress when TRUE.}
}
\description{
Check a single dependency
}
\keyword{internal}
