% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selftest.R
\name{check_example_tree}
\alias{check_example_tree}
\title{Check example tree loading and structure}
\usage{
check_example_tree(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Report progress when TRUE.}
}
\description{
Check example tree loading and structure
}
\keyword{internal}
