% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selftest.R
\name{check_monophyly_logic}
\alias{check_monophyly_logic}
\title{Check monophyly logic with known cases}
\usage{
check_monophyly_logic(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Report progress when TRUE.}
}
\description{
Check monophyly logic with known cases
}
\keyword{internal}
