% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selftest.R
\name{check_taxonomy_extraction}
\alias{check_taxonomy_extraction}
\title{Check taxonomy extraction from example tree}
\usage{
check_taxonomy_extraction(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Report progress when TRUE.}
}
\description{
Check taxonomy extraction from example tree
}
\keyword{internal}
