% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrupt.R
\name{cleanup_on_interrupt}
\alias{cleanup_on_interrupt}
\title{Clean up resources on interruption}
\usage{
cleanup_on_interrupt()
}
\description{
Closes open connections, removes incomplete temp files.
}
\keyword{internal}
