% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-smart.R
\name{compute_legend_layout}
\alias{compute_legend_layout}
\title{Compute legend row/column layout}
\usage{
compute_legend_layout(n_groups, position, nrow, ncol)
}
\arguments{
\item{n_groups}{Number of groups}

\item{position}{Position ("right", "left", "top", "bottom", "inside")}

\item{nrow}{User-specified rows (can be NULL)}

\item{ncol}{User-specified columns (can be NULL)}
}
\value{
list(nrow = integer, ncol = integer)
}
\description{
Compute legend row/column layout
}
\keyword{internal}
