% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescale.R
\name{compute_time_breaks}
\alias{compute_time_breaks}
\title{Compute adaptive time breaks for x-axis}
\usage{
compute_time_breaks(x_min, x_max)
}
\arguments{
\item{x_min}{Minimum x value (negative, in Ma)}

\item{x_max}{Maximum x value (usually 0)}
}
\value{
List with breaks, labels, unit_label
}
\description{
Compute adaptive time breaks for x-axis
}
\keyword{internal}
