% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{convert_unit}
\alias{convert_unit}
\title{Convert time units (Ga -> Ma)}
\usage{
convert_unit(tree, unit)
}
\arguments{
\item{tree}{phylo object}

\item{unit}{Time unit "Ga" or "Ma"}
}
\value{
Modified phylo object (copy, original unchanged)
}
\description{
Convert time units (Ga -> Ma)
}
\keyword{internal}
