% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{.release_log_lock}
\alias{.release_log_lock}
\title{Release the advisory log file lock acquired by \code{.acquire_log_lock()}}
\usage{
.release_log_lock()
}
\description{
Release the advisory log file lock acquired by \code{.acquire_log_lock()}
}
\keyword{internal}
