% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{format_timestamp}
\alias{format_timestamp}
\title{Format timestamp in ISO 8601 with milliseconds and timezone offset}
\usage{
format_timestamp()
}
\description{
Format timestamp in ISO 8601 with milliseconds and timezone offset
}
\keyword{internal}
