% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrupt.R
\name{get_progress_summary}
\alias{get_progress_summary}
\title{Get current progress summary}
\usage{
get_progress_summary()
}
\value{
Character string with progress info.
}
\description{
Get current progress summary
}
\keyword{internal}
