% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{get_rank_name}
\alias{get_rank_name}
\title{Get display name for a taxonomic rank}
\usage{
get_rank_name(rank)
}
\arguments{
\item{rank}{Taxonomic rank (abbreviation or full name)}
}
\value{
Display name string
}
\description{
Get display name for a taxonomic rank
}
\keyword{internal}
