% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-input.R
\name{get_supported_extensions}
\alias{get_supported_extensions}
\title{Get supported file extensions}
\usage{
get_supported_extensions()
}
\value{
Named list of supported extensions by category.
}
\description{
Get supported file extensions
}
