% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{get_version_string}
\alias{get_version_string}
\title{Get package version with git hash if available}
\usage{
get_version_string()
}
\description{
Get package version with git hash if available
}
\keyword{internal}
