% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{init_steps}
\alias{init_steps}
\title{Initialize step counter for progress tracking}
\usage{
init_steps(total)
}
\arguments{
\item{total}{Integer. Total number of steps.}
}
\description{
Initialize step counter for progress tracking
}
\keyword{internal}
