% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_critical}
\alias{log_critical}
\title{Log a CRITICAL message}
\usage{
log_critical(..., .module = NULL)
}
\arguments{
\item{...}{Message components (passed to sprintf if multiple).}

\item{.module}{Character. Optional module/function tag (§15). Default: NULL.}
}
\description{
Log a CRITICAL message
}
\keyword{internal}
