% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_keyvalue}
\alias{log_keyvalue}
\title{Log a key-value pair}
\usage{
log_keyvalue(key, value, level = "INFO")
}
\arguments{
\item{key}{Character. Key name.}

\item{value}{Any. Value to display.}

\item{level}{Character. Log level. Default: "INFO".}
}
\description{
Log a key-value pair
}
\keyword{internal}
