% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_progress}
\alias{log_progress}
\title{Log a progress indicator}
\usage{
log_progress(current, total, item = "")
}
\arguments{
\item{current}{Integer. Current progress.}

\item{total}{Integer. Total items.}

\item{item}{Character. Description of current item.}
}
\description{
Log a progress indicator
}
\keyword{internal}
