% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_section}
\alias{log_section}
\title{Print a formatted section header}
\usage{
log_section(title)
}
\arguments{
\item{title}{Character. Section title.}
}
\description{
Print a formatted section header
}
\keyword{internal}
