% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_subsection}
\alias{log_subsection}
\title{Print a formatted subsection header}
\usage{
log_subsection(title)
}
\arguments{
\item{title}{Character. Subsection title.}
}
\description{
Print a formatted subsection header
}
\keyword{internal}
