% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_table}
\alias{log_table}
\title{Print a formatted table}
\usage{
log_table(data, title = NULL)
}
\arguments{
\item{data}{Named list or data.frame to display.}

\item{title}{Character. Optional table title.}
}
\description{
Print a formatted table
}
\keyword{internal}
