% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_warning}
\alias{log_warning}
\title{Log a WARNING message}
\usage{
log_warning(..., .module = NULL)
}
\arguments{
\item{...}{Message components (passed to sprintf if multiple).}

\item{.module}{Character. Optional module/function tag (§15). Default: NULL.}
}
\description{
Log a WARNING message
}
\keyword{internal}
