% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{next_step}
\alias{next_step}
\title{Increment step counter}
\usage{
next_step(step_name = NULL)
}
\arguments{
\item{step_name}{Character. Description of the step.}
}
\description{
Increment step counter
}
\keyword{internal}
