% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{normalize_newlines}
\alias{normalize_newlines}
\title{Normalize line endings in text}
\usage{
normalize_newlines(text)
}
\arguments{
\item{text}{Character vector or single string.}
}
\value{
Character with normalized line endings.
}
\description{
Converts all line endings to Unix style (LF).
}
\keyword{internal}
