% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{normalize_rank}
\alias{normalize_rank}
\title{Normalize rank abbreviation to full name}
\usage{
normalize_rank(rank)
}
\arguments{
\item{rank}{Rank code or full name ("k","d","p","c","o","f","g","s","ss" or "kingdom","domain","phylum",...)}
}
\value{
Standardized full name string
}
\description{
Normalize rank abbreviation to full name
}
\keyword{internal}
