% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-taxonomy.R
\name{parse_custom_rank}
\alias{parse_custom_rank}
\title{Parse custom rank format labels (Format A wrapper)}
\usage{
parse_custom_rank(labels, levels = NULL, delimiter_mode = "reverse")
}
\arguments{
\item{labels}{Character vector of tip labels}

\item{levels}{List with codes and names vectors.}

\item{delimiter_mode}{Character. Embedded parsing strategy.}
}
\value{
data.frame with taxonomy columns
}
\description{
Parse custom rank format labels (Format A wrapper)
}
\keyword{internal}
