% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-taxonomy.R
\name{parse_custom_regex}
\alias{parse_custom_regex}
\title{Parse custom regex format labels}
\usage{
parse_custom_regex(labels, rank_patterns)
}
\arguments{
\item{labels}{Character vector of tip labels}

\item{rank_patterns}{Named list of regex patterns}
}
\value{
data.frame
}
\description{
Parse custom regex format labels
}
\keyword{internal}
