% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-taxonomy.R
\name{parse_gtdb}
\alias{parse_gtdb}
\title{Parse GTDB format labels (wrapper for parse_semicolon_delimited)}
\usage{
parse_gtdb(labels, levels = NULL, sep = ";")
}
\arguments{
\item{labels}{Character vector of tip labels}

\item{levels}{List with codes and names vectors.}

\item{sep}{Character. Separator between taxonomy ranks. Default: \code{";"}.}
}
\value{
data.frame with taxonomy columns
}
\description{
Parse GTDB format labels (wrapper for parse_semicolon_delimited)
}
\keyword{internal}
