% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-taxonomy.R
\name{parse_silva}
\alias{parse_silva}
\title{Parse Silva format labels}
\usage{
parse_silva(labels)
}
\arguments{
\item{labels}{Character vector of tip labels}
}
\value{
data.frame with columns: domain, phylum, class, order, family, genus, species
}
\description{
Parse Silva format labels
}
\keyword{internal}
