% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{print.rclade_options}
\alias{print.rclade_options}
\title{Print method for rclade_options}
\usage{
\method{print}{rclade_options}(x, ...)
}
\arguments{
\item{x}{A \code{rclade_options} object.}

\item{...}{Ignored.}
}
\value{
Returns the input object \code{x} invisibly, after printing its
key-value pairs to the console.
}
\description{
Print method for rclade_options
}
\keyword{internal}
