% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{RCLADE_HADEAN_END}
\alias{RCLADE_HADEAN_END}
\title{End of the Hadean eon / start of the Archean eon (Ma, ICS 2023/02)}
\usage{
RCLADE_HADEAN_END
}
\description{
End of the Hadean eon / start of the Archean eon (Ma, ICS 2023/02)
}
\keyword{internal}
