% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{RCLADE_X_MARGIN_FACTOR}
\alias{RCLADE_X_MARGIN_FACTOR}
\title{Default margin factor for x-axis range extension}
\usage{
RCLADE_X_MARGIN_FACTOR
}
\description{
Extends the x-axis range by 5\% beyond the tree depth to prevent
tip labels from being clipped at the plot boundary.
Source: ICS chronostratigraphic chart convention for visual padding.
}
\keyword{internal}
