% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-mrca.R
\name{sort_by_depth}
\alias{sort_by_depth}
\title{Sort groups by MRCA node depth (deepest first)}
\usage{
sort_by_depth(groups, mrca_map, tree)
}
\arguments{
\item{groups}{Character vector of group names}

\item{mrca_map}{Output of compute_mrca_map()}

\item{tree}{phylo object}
}
\value{
Sorted character vector
}
\description{
Sort groups by MRCA node depth (deepest first)
}
\keyword{internal}
