% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-deep.R
\name{summarize_multi_trees}
\alias{summarize_multi_trees}
\title{Summarize multiple trees in a file}
\usage{
summarize_multi_trees(trees, filepath)
}
\arguments{
\item{trees}{multiPhylo object.}

\item{filepath}{Character. Source file path.}
}
\description{
Prints a summary of each tree in a multiPhylo object before
raising the multi-tree error.
}
\keyword{internal}
