% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-timetree.R
\name{summarize_timetree}
\alias{summarize_timetree}
\title{Print a summary of a Rclade timetree plot}
\usage{
summarize_timetree(p)
}
\arguments{
\item{p}{ggplot object returned by plot_timetree() (with rclade_info attribute)}
}
\value{
Invisibly returns the info list
}
\description{
Print a summary of a Rclade timetree plot
}
