% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{timer_start}
\alias{timer_start}
\title{Start a timer for performance measurement}
\usage{
timer_start(name)
}
\arguments{
\item{name}{Character. Timer name.}
}
\description{
Start a timer for performance measurement
}
\keyword{internal}
