% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrupt.R
\name{update_progress}
\alias{update_progress}
\title{Update progress counter}
\usage{
update_progress(processed = NULL, current = NULL)
}
\arguments{
\item{processed}{Integer. Number processed so far.}

\item{current}{Character. Description of current item.}
}
\description{
Update progress counter
}
\keyword{internal}
