% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-mrca.R
\name{validate_collapse_plan}
\alias{validate_collapse_plan}
\title{Detect nesting conflicts in a collapse plan}
\usage{
validate_collapse_plan(mrca_map, tree)
}
\arguments{
\item{mrca_map}{Output of compute_mrca_map()}

\item{tree}{phylo object}
}
\value{
Invisibly returns warning message vector
}
\description{
Detect nesting conflicts in a collapse plan
}
\keyword{internal}
