% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-deep.R
\name{validate_fasta_content}
\alias{validate_fasta_content}
\title{Validate FASTA content}
\usage{
validate_fasta_content(lines, filepath, result, check_alignment = FALSE)
}
\arguments{
\item{lines}{Character vector of file lines.}

\item{filepath}{Character. File path for messages.}

\item{result}{List to populate.}

\item{check_alignment}{Logical. Whether to check sequence length consistency.}
}
\description{
Validate FASTA content
}
\keyword{internal}
