% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-deep.R
\name{validate_fastq_content}
\alias{validate_fastq_content}
\title{Validate FASTQ content}
\usage{
validate_fastq_content(lines, filepath, result)
}
\description{
Validate FASTQ content
}
\keyword{internal}
