% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-input.R
\name{validate_tree_structure}
\alias{validate_tree_structure}
\title{Validate tree object structure}
\usage{
validate_tree_structure(tree)
}
\arguments{
\item{tree}{phylo object to validate.}
}
\value{
Logical. TRUE if valid.
}
\description{
Validate tree object structure
}
\keyword{internal}
