% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{set_log_file}
\alias{set_log_file}
\title{Set log file for dual output}
\usage{
set_log_file(filepath)
}
\arguments{
\item{filepath}{Character. Path to log file. NULL to disable.}
}
\description{
Set log file for dual output
}
