% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{set_log_level}
\alias{set_log_level}
\title{Set log level}
\usage{
set_log_level(level)
}
\arguments{
\item{level}{Character. One of "DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL".}
}
\value{
No return value, called for side effects (sets the package-level
log threshold).
}
\description{
Set log level
}
